requirements.txt
aiohttp==3.9.0
aiokafka==0.10.0
async-timeout==4.0.3
ccxt==4.2.86
ccxtpro==0.10.54
cryptography==42.0.2
dotenv==0.0.5
pydantic==2.6.4
python-dotenv==1.0.1
PyYaml==6.0.1
uvloop==0.19.0
websockets==12.0
fastapi==0.110.0
uvicorn[standard]==0.27.0
prometheus-client==0.20.0
pytest==8.1.1
pytest-asyncio==0.23.6
mypy==1.9.0
